Implemented Access Control

+++ app/Http/Controllers/PostsController.php
@@ -59,6 +59,10 @@ class PostsController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id)
+            return redirect('posts')->with('error', 'Unauthorized page');
+
         return view('postpages/show')->with('post', $post);
     }
 
@@ -105,6 +109,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id)
+            return redirect('posts')->with('error', 'Unauthorized page');
+
         $post->delete();
 
         return redirect('posts')->with('success', 'Post Removed');
